Fix a bug in doeclim.jl for the calculation of the 'Cdoe' variable.
QUICK VERSION: Missing parenthesis

LONG VERSION: In the fortran version of doeclim, the Cdoe variable is a
2x2 array. The four values for Cdoe are calculated separately, and then
each term is multiplied as such:

Cdoe = Cdoe * (deltat**2/12.)

In the julia Mimi version of doeclim, each term in Cdoe was calculated
and then multiplied by (deltat**2/12.) all at once. The first part of
each equation was not enclosed in parenthesis, so only the last terms
after an addition sign were multiplied by (deltat**2/12.). This fix has
the Mimi and fortran versions of doeclim matching on output variables
and parameters for all timesteps.

Conflicts:
	src/doeclim.jl